templates/subsidy-consultant/pyproject.toml
[project]
name = "subsidy-consultant"
version = "1.0.0"
description = "AI-powered subsidy/grant consulting assistant for Japanese SMEs. Uses Azure AI Foundry to match grants, draft applications, and score proposals."
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = { text = "MIT" }

dependencies = [
    "crewai>=1.9,<2.0",
    "crewai-tools>=0.36",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "openai>=1.50",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
